Create new project before attaching it
When attaching a directory to a project during startup, saving is
disallowed, so the attached directory only has a `workspace.xml` in its
`.idea` when the attach is attempted.  Attaching requires the
`.idea/*.iml` Module file, so the attaching fails, saying the directory
is an unsupported format.

Experimentation showed that manually attaching the directory also did
not work, but opening the directory in a separate window, then opening
and attaching it again would make the directory have the full project
files.  To mimic this manual process:

1. The internals of `doOpenProject` are copied
2. A save of the project files is forced, bypassing the normal "startup"
   save blocker
3. The project is attached to the root project.

Tested to work when upgrading from 10.4.0 to 10.5.0-pre in Rubymine when
no project was already open.