Remove project converts
Project converters are broken in 2019.2 because the DialogWrapper
they try to launch isn't being launched in the Event Dispacht Thread
(EDT) (https://youtrack.jetbrains.com/issue/IDEA-218071).

As this is bug in code wholey controlled by upstream, the only option
is to remove the project converters.  The work-around is for users
to re-import projects when a project layout update is necessary.