Fix ConcurrentModificationException in Structure View
Fixes #1488

Java 9 fixed a bug (https://bugs.openjdk.java.net/browse/JDK-8071667) in
`HashMap` where `computeIfAbsent` did not check for concurrent modifications,
and it turns out that `TreeElementList` was using concurrent modifications,
so it was now broke.

Fixed by use `get` or `put` if it is absent, so that `putNew` can ensure that
the `CallDefinition` is in the `TreeElementList` before it is added to the
`MutableList<TreeElement>`, which was the original reason why there was a `put`
inside of `computeIfAbsent`, which would have `put` when the function returned
anyway.