Don't use parentheses in Elixir version eval
Fixes #1597

GeneralCommandLine's escaping for Windows can't handle the parentheses
in a way that both works for the Windows shell and Elixir running the
code the shell hands off.  Removing the parentheses leaves runnable
code even if it is no longer formatted.  Fixes "Unknown Version" naming
for Elixir SDKs and the "Probably SDK installed in ... is corrupt"
dialog from appearing.