Allow arity mismatch to resolve as invalid result
Previously, the arity of a function or macro could only disagree with
the definition and resolve if

1. The arity of the call was in the arity range of the defintion and the
`ResolveResult` was marked as a `validResult`.
2. The code is marked as `incomplete` (because the IDE detects typing,
etc) and the arity mismatches and the `ResolveResult` is marked as an
invalid result.

By marking all arity mismatches, regardless of incompleteness of the
code as an invalid result, Go To Definition and Show Parameters works
when the the arity is incorrect.  This males Show Parameters work while
typing out a call without the correct number of commas and allows
jumping to the definition while typing out the call too.