Vendor JInterface 1.11
Fixes #1784

The JInterface on Maven Central has stopped being updated because the OTP Team didn't actually maintain it, that was Basho, and Basho is gone now.  This version of JInterface, 1.11, is from Erlang 23.0.4 from Homebrew, but with the formula edited (`brew edit erlang`) to add `--with-java` and then built with `brew install erlang --build-from-source`.