Connect compiled stubs to decompiled source by name/arity
Previously, when the decompiler didn't use the Docs chunk, it was guaranteed that the PsiCompiled stubs would correlate, in-order, with the decompiled source call definitions, and so mirrors could be set by matching up the list of them in order.  Since the Docs doesn't have to correspond to and doesn't correspond to the binary precisely for some BEAMs, most importantly, `Elixir.Kernel.beam`, the PsiCompiled stub and decompiled source is now matched by name and arity.   This means some mirrors are missed, but no errors are produced.