Use ~S for docs from Docs chunk
The docs from the docs chunk may contain interpolation in the code samples or `#{` in regex examples, but these should not be treated as an interpolation start as the docs format does not support interpolation.  Anything that looks like interpolation in the docs text was actually escaped in the original docs, so also escape it here by using `S"""`, which turns off interpolation.