handleInLastState unexcepted atom starts instead of marking as BAD_CHARACTER
Fixes #1888
Fixes #1889

For a reason I haven't been able to explain, the LexerEditorHighlighter stops working after `:` is typed at the start of an atom in certain situations, such as before `)` inside a function call, like when adding an argument.  In this case, the old version of the lexer would mark `)` as a BAD_CHARACTER and continue to do so until an expected atom start of [a-zA-Z_], `'`, `"`, or an operator occurred.  Now, if an invalid atom start is matched, the `ATOM_START` state ends and the previous state handles the text, which in the function case mean `)` is parsed as `CLOSING_PARENTHESIS`.