Use VirtualFile#fileType instead of EEx Type::INSTANCE when looking up extensions
Fixes #1869

Since LEEx file `Type` is a a subclass of EEx's file `Type`, it calls `templateDataFileTypeSet` in EEx's `Type`, but `templateDataFileTypeSet` uses `INSTANCE` from EEx.  By using the `VirtualFile#fileType` instead, it will properly be EEx or LEEx based on the actual file extension and then it can be used to strip that off and find the DataTemplateLanguage, such as `HTMLLanguage` for `.html.leex`.