Reimplment Module references
Instead of references for only the outermost QualifiableAlias, there is a reference for each right-most Alias at a given position, so instead of there only be a reference to `App.Context.Schema` in `App.Context.Schema`, there is now a reference to `App` in the `App` prefix, a reference to `App.Context` in `Context` in `App.Context`, and a reference to `App.Context.Schema` in `Schema` in `App.Context.Schema`.  Not only is this more useful, being able to jump to parent namespaces, but it fixed some of the capability issues with Go To Definition in the 2020 line of IDEs.  This approach of using `getRangeInElement` to target the range of the right-most Alias, while the element was still the parent that contained, but did not go beyond the Alias, was tried after having references only on `Alias`es and not `QualifiedAlias`es did not fix completion issues.  It was the while debugging Go To Declaration actions and noticing they were sensitive to the range in element AND the [docs for PsiReference#getRangeInElement](https://github.com/JetBrains/intellij-community/blob/f66e9644fa683fba22f4ce9e30c037720f745989/platform/core-api/src/com/intellij/psi/PsiReference.java#L49-L62) that I realized that the Go To Declaration and Completion has a hidden requirement that References for things that behave like namespaces have to work this way.