Fake built-in types by decompiling in :erlang.beam
For type reference resolution of built-in types to work, they need to be defined somewhere.  IntelliJ Erlang does this by looping usages back on themselves, but this leads to Find Usages not working for built-in types since each usage is a distinct declaration.  By instead defining the types in decompiled :erlang.beam (even though they aren't actually defined there), there is a shared location for all reference to point to and then check for Find Usages.