Fix resolvePrimaryArity for `|> case do`
resolvedPrimaryArity did not add both the doBlock's arity and the arity from the `|>` at the same time.  If the pipe was there, the `do` was ignored because the pre-existing `resolvePrimaryArity` was not used if the normal primary arity of the call was zero, which is the case for `|> case do`.