Support arity intervals for unquote_splicing in parameters
Functions defined with unquote_splicing, such as `Ecto.Schema.__schema/2`:

```elixir
for clauses <- Ecto.Schema.__schema__(fields, field_sources, assocs, embeds),
    {args, body} <- clauses do
  def __schema__(unquote_splicing(args)), do: unquote(body)
end
```

Need to have their arity not be the number of PsiElements in the parentheses.  Any call to `unquote_splicing(...)` can end up have 0 to infinite parameters, so it means when one is saw, the range of `minimum...maximum` should change to an open interval of `minimum...`.  This required changing `IntRange resolvedFinalArityChange()` to `ArityInterval resolvedFinalArityInterval()` on all `Call`s, which was a large change.  It also meant changing a lot of ArityRange types to ArityInterval, and NameArityRange to NameArityInterval, which influenced the variable names.

Since all Calls support ArityIntervals now and not just special forms and Ecto DSLs, exportArity is changed to always state the ResolveState, so that the special form changes can be integrated for all callers.

The actual implementation of CallImpl.resolvedFinalArityRange is changes to fold over the ArityInterval:
* Normal arguments increase the minimum and maximum.
* Default arguments increase only the maximum.
* unquote_splicing changes the maximum to null to indicate the interval is half open.