Don't check following siblings of modulars if entrance is a direct child
Prevent StackOverflow when trying to resolve embed_template when more than one appears in the same module.

In general, if the entrance is a child of modular then it can only be defined by a previous sibling, usually an `import` or `use`, but if the entrance is descendant of a child, then it child then it may be a call to a function or macro defined in the modular to following siblings of the entrance ancestor child needs to be checked if the entrance is a forward-call to a later declared function or macro.