Drop support for Elixir <= 1.6
Continuing support for Elixir <= 1.6 required special handling of the language level to support differences in precedence and operators.  Removing the language level tracking allows dropping the Level and FilePropertyPusher.  It also eliminates the need for tests for setup the SDK since it was only needed to get the appropriate Level.  This makes the tests run in 45 seconds instead of 7 minutes.