The parent argument to AccumulatorContinue.childExpressionsFoldWhile should be this and not parent
When converting to an extension function I left `parent` in place because the argument is called `parent`, but since it is an extension function that value because `this.parent` when it really should have been `this`.  Using `this.parent` meant it would ask for the parent's children and keep looping back to `this`.