Setting name on a call definition clause sets the name on the head instead
Fixes #2094
Fixes #2917

The JetBrains API both renames the usage references AND the original element where the rename refactoring was started, but when started on a call definition clause the `functionNameElement` is the `def`, `defp`, `defmacro`, or `defmacrop` and so both the name of the call was changed, but the `def*` was also changed to the new name.  Now, if `setName` is called on a definition, the head is renamed instead.