Keep looking for variants above if and unless
Fixes #2751

`if` and `unless` used to call `keepProcessing()` after checking if the child calls when resolving or searching for variants, but Variants returned `false` always from `keepProcesssing()`.

This stops the variant search from terminating inside of an `if` or `unless` when it wouldn't outside of one as was reported in #2751.